refactor: extract api-jobs data types into homeboy-api-jobs-contract
The api_jobs job model (Job, JobStatus, JobEvent, JobEventKind,
RunnerJobLogSnapshot, JobClaimMetadata, and daemon/lease diagnostics -- 17
pure serde types) is consumed cross-crate heavily (JobStatus 42x, JobEvent
15x, Job 15x) but lived in core alongside the 104k-LOC job store.

Split:
- api_jobs/types.rs moves DOWN into new leaf crate homeboy-api-jobs-contract.
  Its deps are all leaves now: SourceSnapshot (homeboy-source-snapshot-
  contract), PathMaterializationPlan (homeboy-lab-contract), uuid, serde.
- The 2 pure metadata structs it needs (JobArtifactMetadata,
  RunnerJobLifecycleMetadata) moved out of the behavior-heavy remote_runner.rs
  into the contract's metadata module; core's remote_runner re-exports them so
  its From<RunnerJobLifecycleMetadata> impl and other refs are unchanged.
- Widened JobStatus::is_terminal pub(super)->pub (core store/persistence/
  remote_runner call it cross-crate now).

Everything else stays in core: store.rs, persistence.rs, remote_runner
behavior, RunnerJobPreparationProvider hook, agent_task_terminal_recovery,
summary. Core re-exports the contract as api_jobs::types, so all consumer
paths (agents, lab-runner, cli) resolve unchanged.

Depends on the source-snapshot extraction (same branch) since Job embeds
SourceSnapshot.

Verified: contract compiles, core --lib clean, 98 api_jobs tests pass,
agents/lab-runner/cli --lib clean, fmt clean.
