+++ crates/contracts/homeboy-api-jobs-contract/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "homeboy-api-jobs-contract"
+version = "0.1.0"
+edition = "2021"
+license = "MIT"
+authors = ["Chris Huber <chubes@extrachill.com>"]
+description = "Pure serializable job model contract types for homeboy (Job/JobStatus/JobEvent)"
+repository = "https://github.com/Extra-Chill/homeboy"
+publish = false
+
+[lib]
+name = "homeboy_api_jobs_contract"
+path = "src/lib.rs"
+
+[dependencies]
+homeboy-lab-contract = { version = "0.1.0", path = "../homeboy-lab-contract" }
+homeboy-source-snapshot-contract = { version = "0.1.0", path = "../homeboy-source-snapshot-contract" }
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
+uuid = { version = "1.11", features = ["v4", "v5", "serde"] }
